test: share arc click e2e harness
